+++ changelog.md
@@ -1,4 +1,15 @@
 # Changelog
 
-**./src/generate-database-entries.js - generateOther**
-* Populate 'otherAvailability' column.
+**./src/read-input-data.js - coordinateData**
+* Uncommented 'callNameData' task.
+
+---
+
+**./src/generate-database-entries.js - prepareBaseData**
+* Uncommented 'localName' variable.
+
+---
+
+**./src/generation/gen-data.js - defineBaseObject**
+* 'name' is now declared as a blank string.
+	* No longer using a placeholder.